refactor: extract repository downloads
