+++ vuln_checker/main.py
@@ -1,25 +1,16 @@
 import itertools
 import json
 import logging
-import os
-import shutil
 from collections import defaultdict
 from operator import itemgetter
 from pathlib import Path
 from typing import Iterable, TypedDict, Literal
 
 import networkx as nx
-import yaml
-from git import Repo, RemoteProgress
-from tqdm import tqdm
 
+from vuln_checker.config import FileConfig
 from vuln_checker.converter import convert_vulnerabilities
-
-DEFAULT_DATABASES_PATH = Path.cwd() / "vuln-list"
-VULNS_PATH = Path.cwd() / "vulnerabilities"
-
-VULN_LIST_REPO = "https://github.com/aquasecurity/vuln-list.git"
-GLAD_DATABASE_REPO = "https://gitlab.com/gitlab-org/security-products/gemnasium-db.git"
+from vuln_checker.load import prepare_databases
 
 
 def check_databases_folder(databases_path: Path):
@@ -31,64 +22,6 @@ def set_to_list(obj):
     if isinstance(obj, set):
         return list(obj)
     raise TypeError
-
-
-class Progress(RemoteProgress):
-    def __init__(self):
-        super().__init__()
-        self.pbar = tqdm()
-
-    def update(self, op_code, cur_count, max_count=None, message=''):
-        self.pbar.total = max_count
-        self.pbar.n = cur_count
-        self.pbar.refresh()
-
-
-def prepare_databases():
-    if os.path.exists(DEFAULT_DATABASES_PATH):
-        logging.info("The databases are already loaded")
-        return
-
-    logging.info("Clone `vuln_list` repo")
-    Repo.clone_from(
-        VULN_LIST_REPO,
-        DEFAULT_DATABASES_PATH,
-        progress=Progress(),
-        single_branch=True,
-        depth=1
-    )
-
-    glad_database_path = os.path.join(DEFAULT_DATABASES_PATH, "glad")
-
-    logging.info("Remove `glad` database from `vuln_list`")
-    if os.path.exists(glad_database_path):
-        shutil.rmtree(glad_database_path)
-
-    logging.info("Clone `GitLab Advisory Database` repo")
-    Repo.clone_from(
-        GLAD_DATABASE_REPO,
-        glad_database_path,
-        progress=Progress(),
-        single_branch=True,
-        depth=1
-    )
-
-    logging.info("Convert files .yml in .json in the `glad` database")
-    glad_path = Path(glad_database_path)
-    packages = ("conan", "gem", "go", "maven", "npm", "nuget", "packagist", "pypi")
-    for directory in glad_path.iterdir():
-        if directory.name not in packages:
-            if directory.is_dir():
-                shutil.rmtree(directory.absolute())
-            else:
-                directory.unlink()
-            continue
-        for p in tqdm((glad_path / directory).rglob("*.yml"), desc=f"Convert `{directory.name}` package"):
-            json_advisory = p.with_suffix(".json")
-            with p.open("r") as yaml_file, json_advisory.open("w") as json_file:
-                advisory = yaml.safe_load(yaml_file)
-                json.dump(advisory, json_file)
-            p.unlink()
 
 
 priorities = frozenset([
@@ -114,11 +47,11 @@ def key_by_priority(keys: Iterable[str], default) -> tuple[str, set[str]]:
     return key, aliases
 
 
-def create_map_of_vulnerabilities() -> VulnMap:
+def create_map_of_vulnerabilities(databases_path: Path, vulnerabilities_path: Path) -> VulnMap:
     vulns: VulnMap = {}
 
-    for database_name, model in convert_vulnerabilities(DEFAULT_DATABASES_PATH):
-        database_path = VULNS_PATH / database_name
+    for database_name, model in convert_vulnerabilities(databases_path):
+        database_path = vulnerabilities_path / database_name
         database_path.mkdir(exist_ok=True)
         vulnerability_file = database_path / (model.id + '.json')
         with vulnerability_file.open('w') as f:
@@ -175,9 +108,9 @@ def combine_vulnerabilities(vulns: VulnMap) -> VulnMap:
     return traverse_graph_and_combine_vulnerabilities(graph)
 
 
-def dump_vulns(vulns: VulnMap):
-    with (VULNS_PATH / "combined.json").open("w") as f:
-        json.dump(vulns, f, default=set_to_list, indent=2)
+def dump_vulns(vulnerabilities: VulnMap, vulnerabilities_path: Path):
+    with (vulnerabilities_path / "combined.json").open("w") as f:
+        json.dump(vulnerabilities, f, default=set_to_list, indent=2)
 
 
 Advisory = TypedDict(
@@ -240,7 +173,7 @@ def find_different_advisories(vulns: VulnMap) -> DifferentAdvisories:
     return different_advisories
 
 
-def report_to_file(advisories: DifferentAdvisories):
+def report_to_file(advisories: DifferentAdvisories, vulnerabilities_path: Path):
     message = ""
     for k, groups in advisories.items():
         for group in groups:
@@ -249,22 +182,26 @@ def report_to_file(advisories: DifferentAdvisories):
                 message += "\t" + "\n\t".join(advisory_to_short_info(adv) for adv in advisories) + "\n"
             message += "-" * 20 + "\n"
 
-    (VULNS_PATH / "report.txt").write_text(message)
+    (vulnerabilities_path / "report.txt").write_text(message)
 
 
 def main():
     logging.basicConfig(level=logging.INFO)
 
-    prepare_databases()
-    VULNS_PATH.mkdir(exist_ok=True)
+    prepare_databases(FileConfig.DEFAULT_DATABASES_PATH)
+    FileConfig.VULNERABILITIES_PATH.mkdir(exist_ok=True)
 
-    vulnerabilities = create_map_of_vulnerabilities()
+    vulnerabilities = create_map_of_vulnerabilities(
+        FileConfig.DEFAULT_DATABASES_PATH,
+        FileConfig.VULNERABILITIES_PATH
+    )
+
     combined_vulnerabilities = combine_vulnerabilities(vulnerabilities)
-    dump_vulns(combined_vulnerabilities)
+    dump_vulns(combined_vulnerabilities, FileConfig.VULNERABILITIES_PATH)
 
     advisories = find_different_advisories(combined_vulnerabilities)
 
-    report_to_file(advisories)
+    report_to_file(advisories, FileConfig.VULNERABILITIES_PATH)
 
 
 if __name__ == "__main__":
